Backend/requirements.txt
channels==4.2.0
Django==5.1.5
langchain==0.3.15
langchain_community==0.3.15
langchain_core==0.3.31
langchain_openai==0.3.2
langchain_teddynote==0.3.41
psutil==6.1.1
python-dotenv==1.0.1
Requests==2.32.3
xmltodict==0.14.2
